Modelling/colab_requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
blinker==1.7.0
blis==0.7.11
catalogue==2.0.10
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
confection==0.1.3
cymem==2.0.8
datasets==2.11.0
dill==0.3.6
filelock==3.13.1
Flask==2.3.2
frozenlist==1.4.0
fsspec==2023.10.0
huggingface-hub==0.19.0
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
langcodes==3.3.0
loralib==0.1.2
MarkupSafe==2.1.3
multidict==6.0.4
murmurhash==1.0.10
nltk==3.8.1
numpy==1.24.2
packaging==23.2
pandas==2.0.3
pathy==0.10.3
Pillow==9.5.0
preshed==3.0.9
pyarrow==14.0.1
pydantic==1.10.13
pyDeprecate==0.3.2
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.3.23
requests==2.28.2
responses==0.18.0
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
smart-open==6.4.0
spacy==3.5.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
thinc==8.1.12
threadpoolctl==3.2.0
tokenizers==0.13.3
torchmetrics==0.8.1
torchtext==0.9.0
tqdm==4.65.0
transformers==4.27.4
typer==0.7.0
typing-extensions==4.8.0
tzdata==2023.3
urllib3==1.26.18
wasabi==1.1.2
werkzeug==3.0.1
xxhash==3.4.1
yarl==1.9.2
zipp==3.17.0